Move logic for legacy installs to dedicated module
